requirements.txt
pandas>=0.22.0
matplotlib>=2.1.0
PyYAML>=3.12
requests>=2.19.1
